slight adjustment to golden file section of README
Come to think of it, maybe it's not so great after all that JSON files in `task/testdata/` directory are used both as input data and as golden data, since updating the expected output can also change the input for some tests. But I still think I prefer this to having a bunch of duplicate files (e.g. `cancelled.input` and `cancelled.golden`).